backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
pyjwt==2.8.0
bcrypt==4.1.3
python-multipart==0.0.9
aiofiles==23.2.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-ollama==0.1.1
langchain-community==0.2.1
faiss-cpu==1.8.0
pypdf==4.2.0
